requirements.txt
openai>=1.6.0
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.7.0
PySide6>=6.6.0
fastapi>=0.110.0
uvicorn>=0.27.0
requests>=2.31.0
markdown>=3.5.0
